removed unused files from before refactoring
